api/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.37
alembic==1.14.1
psycopg2-binary==2.9.10
redis==5.2.1
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.7.1
docker==7.1.0
httpx==0.28.1
python-multipart==0.0.20
minio==7.2.15
pyjwt[crypto]==2.10.1
arq==0.26.1
